+++ messenger/angular.json
@@ -28,7 +28,8 @@
               "src/assets"
             ],
             "styles": [
-              "src/styles.scss"
+              "src/styles.scss",
+              "./node_modules/@ctrl/ngx-emoji-mart/picker.css"
             ],
             "scripts": []
           },
