+++ messenger/src/app/app.component.scss
@@ -1,5 +1,6 @@
 :host {
   display: flex;
   flex-direction: column;
+  justify-content: space-between;
   height: 100%;
 }
