+++ messenger/src/app/modules/chat/chat.component.scss
@@ -1,18 +1,6 @@
 :host {
   display: block;
-  height: 95%;
-}
-
-.chat-box {
-  width: 450px;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
+  height: calc(100% - 120px);
+  width: 90%;
   margin: 0 auto;
-}
-
-.messages-box {
-  height: 92%;
-  overflow-y: auto;
 }
