+++ messenger/src/app/modules/chat/chat.module.ts
@@ -4,20 +4,28 @@ import { ChatComponent } from './chat.component';
 import {RouterModule, Routes} from "@angular/router";
 import {FormsModule} from "@angular/forms";
 import { ChatMessageComponent } from './components/chat-message/chat-message.component';
+import {PickerModule} from "@ctrl/ngx-emoji-mart";
+import { EmojiModule } from '@ctrl/ngx-emoji-mart/ngx-emoji';
+import { ChatRoomComponent } from './components/chat-room/chat-room.component';
 
 const routes: Routes = [
-  {path: '', component: ChatComponent}
+  {path: '', pathMatch: 'full', component: ChatComponent},
+  {path: 'chat/:id', component: ChatRoomComponent},
+
 ];
 
 @NgModule({
   declarations: [
     ChatComponent,
-    ChatMessageComponent
+    ChatMessageComponent,
+    ChatRoomComponent
   ],
   imports: [
     CommonModule,
     RouterModule.forChild(routes),
-    FormsModule
+    FormsModule,
+    PickerModule,
+    EmojiModule
   ]
 })
 export class ChatModule { }
