+++ messenger/src/app/modules/chat/components/chat-message/chat-message.component.html
@@ -1,6 +1,6 @@
-<li [ngClass]="{ 'chat-right': isUser, 'chat-left': !isUser }">
+<li class="chat-message" [ngClass]="{ 'chat-right': isUser, 'chat-left': !isUser }">
   <div class="chat-avatar">
-    <img [src]="message.authorPhoto" (error)="errorHandler($event)">
+    <img class="avatar-img" [src]="message.authorPhoto" (error)="errorHandler($event)">
     <div class="chat-name">{{ message?.author }}</div>
     <div class="chat-hour">{{ message?.timestamp | date:'short' }}</div>
   </div>
