+++ messenger/src/app/modules/chat/components/chat-message/chat-message.component.scss
@@ -1,12 +1,16 @@
-li {
+.chat-message {
+  max-width: 430px;
+  width: 93%;
   &.chat-left, &.chat-right {
     display: flex;
     flex: 1;
     flex-direction: row;
     margin-bottom: 40px;
+    max-width: 430px;
+    width: 93%;
   }
 
-  img {
+  .avatar-img {
     width: 48px;
     height: 48px;
     -webkit-border-radius: 30px;
@@ -56,6 +60,7 @@ li {
     font-weight: 300;
     line-height: 150%;
     position: relative;
+    word-break: break-word;
 
     &:before {
       content: '';
@@ -86,12 +91,12 @@ li {
 }
 
 @media (max-width: 767px) {
-  li.chat-left, li.chat-right {
+  .chat-message.chat-left, .chat-message.chat-right {
     flex-direction: column;
     margin-bottom: 30px;
   }
 
-  li img {
+  .chat-message .avatar-img {
     width: 32px;
     height: 32px;
   }
@@ -132,8 +137,9 @@ li {
     }
   }
 
-  li .chat-text {
+  .chat-message .chat-text {
     font-size: .8rem;
+    margin-left: 15px;
   }
 
 }
