+++ messenger/src/app/modules/chat/components/chat-room/chat-room.component.scss
@@ -1,0 +1,76 @@
+:host {
+  display: block;
+  height: calc(100% - 120px);
+}
+
+.emoji-mart-block {
+  position: absolute;
+  bottom: 67px;
+  left: 0;
+}
+
+.send-wrapper {
+  position: relative;
+}
+
+.chat-box {
+  max-width: 450px;
+  width: 90%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  margin: 0 auto;
+  padding: 10px;
+  background: #cbd0e8;
+  border-radius: 5px;
+}
+
+.messages-box {
+  height: 92%;
+  overflow-y: auto;
+}
+
+.input {
+  padding: 15px 58px 15px 15px;
+  width: 100%;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: #ffffff;
+  border-top: 1px solid white;
+  outline: none;
+}
+
+.send-button {
+  position: absolute;
+  top: 14px;
+  right: 5px;
+  background: none;
+}
+
+.open-emoji-button {
+  position: absolute;
+  top: 13px;
+  right: 34px;
+  background: none;
+}
+
+.messages-box::-webkit-scrollbar,
+.messages-box .autocomplete-items::-webkit-scrollbar {
+  width: 12px;
+}
+
+.messages-box::-webkit-scrollbar-track,
+.messages-box::-webkit-scrollbar-track {
+  box-shadow: inset 0 0 6px rgba(0,0,0,0.3);
+  -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3);
+  border-radius: 10px;
+}
+
+.messages-box::-webkit-scrollbar-thumb,
+.messages-box::-webkit-scrollbar-thumb {
+  border-radius: 10px;
+  box-shadow: inset 0 0 6px rgba(0,0,0,0.5);
+  -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.5);
+}
