+++ messenger/src/app/modules/layout/footer/footer.component.html
@@ -1,1 +1,1 @@
-<p>footer works!</p>
+<footer class="footer-content">Footer</footer>
