+++ messenger/src/app/modules/layout/footer/footer.component.scss
@@ -1,0 +1,10 @@
+:host {
+  display: block;
+  height: 60px;
+  border-top: 1px solid gray;
+}
+
+.footer-content {
+  width: 90%;
+  margin: 0 auto;
+}
