+++ messenger/src/app/modules/layout/header/header.component.html
@@ -1,2 +1,7 @@
-<div>{{user?.displayName}}</div>
-<button (click)="logOut()">Log out</button>
+<header class="header-content">
+  <div class="logo"><img class="logo-image" src="/assets/images/logo.png" alt="logo"></div>
+  <div>
+    <div>{{user?.displayName}}</div>
+    <button *ngIf="user?.displayName" (click)="logOut()">Log out</button>
+  </div>
+</header>
