+++ messenger/src/app/modules/layout/header/header.component.ts
@@ -26,6 +26,7 @@ export class HeaderComponent implements OnInit, OnDestroy {
 
   public logOut() {
     this.authService.logout();
+    this.user = null;
   }
 
 }
