+++ messenger/src/styles.scss
@@ -1,4 +1,6 @@
 /* You can add global styles to this file, and also import other style files */
+@import '~@ctrl/ngx-emoji-mart/picker';
+
 *,
 *::after,
 *::before {
